requirements.txt
streamlit
openai-whisper
transformers
torch
numpy
pydub
scikit-learn
textblob
openai
pydub
SpeechRecognition
